Agregado que el broker otorgador actualice la shm de version. Queda pendiente de aprobacion porque tuve que cambiar los tipos de datos que se transmitian al generar el anillo

+++ Ejercicio 6/v5 dist. varios brokers con anillo y lider/Anillo/inet.h
@@ -67,7 +67,7 @@ typedef struct{
 
 
 typedef struct{
-    int idBroker;
+    uint64_t idBroker;
     int estado;
 }MsgLider_t;
